move this to a django test
